Switched to using a new "EmbeddedDIFx" reusable component via NuGet package. * Simplifies and abstracts "Use DIFx to install a driver" scenarios for reuse,   via a simple 'using' syntax for DIFx. * Eliminates the need to directly manage multiple dependencies for bitness; the   correct version of DIFxAPI is extracted and used at run-time. Also, other ScpDriverInterface cleanup and improvements: * Removed unused references. * Reduced minimum framework version needed to use ScpDriverInterface. * Fixed a casting warning. * Some improved error handling.

+++ ScpDriverInterface/ScpBus.cs
@@ -230,7 +230,7 @@ namespace ScpDriverInterface
 
 					if (NativeMethods.SetupDiGetDeviceInterfaceDetail(deviceInfoSet, ref DeviceInterfaceData, detailDataBuffer, bufferSize, ref bufferSize, ref da))
 					{
-						IntPtr pDevicePathName = detailDataBuffer + 4;
+						IntPtr pDevicePathName = (IntPtr)((Int64)detailDataBuffer + 4);
 
 						path = Marshal.PtrToStringAuto(pDevicePathName).ToUpper(CultureInfo.InvariantCulture);
 						Marshal.FreeHGlobal(detailDataBuffer);
